More feedback if server fails to start
If the server fails to start, we now print the stderr output from Java and
also reject the promise with an appropriate error message.
In particular this will now print some useful information if java was not
found, or vnu.jar was not found, or some such.

If the process exits at some later point, after it had already started, then
we spuriously try to reject the promise as well, but that's no problem since
we already accepted it before, and the first decision sticks.