If JAVA_HOME is not set, use PATH to find the location of the java binary
child_process.spawn will by default use the PATH to resolve the path of a
binary if only its name is given.  We can make use of this for the case when
JAVA_HOME is not set.  Just calling the first java binary on the PATH is
usually more portable then hard-coding /bin/java.