Fix uncaught test error
The finally handler is always called without arguments, so passing "done" as
an argument to that will result in "done" being called with no arguments,
indicating a successful run.  The other scheme is what's used in most of the
other tests: call done with no arguments on success, and with error object
on failure.